Healthcare/healthcare_analytics/requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Machine Learning
xgboost>=1.5.0
tensorflow>=2.8.0  # For LSTM models
shap>=0.40.0
lime>=0.2.0.1

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0  # For PostgreSQL

# Utilities
jupyter>=1.0.0
ipykernel>=6.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.0

# Time Series Analysis
statsmodels>=0.13.0

# Data Processing
pyarrow>=6.0.0
fastparquet>=0.8.0
